Separate AlertList and RouteList
