+++ Linkbus/Linkbus/OnboardingView.swift
@@ -79,15 +79,15 @@ struct InformationContainerView: View {
         .padding(.horizontal)
     }
     else {
-    VStack(alignment: .leading) {
-        InformationDetailView(title: "Lightweight", subTitle: "Fast and lightweight app makes it easier than ever to view the bus schedule.", imageName: "wind")
+        VStack(alignment: .leading) {
+            InformationDetailView(title: "Lightweight", subTitle: "Fast and lightweight app makes it easier than ever to view the bus schedule.", imageName: "wind")
 
-        InformationDetailView(title: "Accurate", subTitle: "Automatically pulls data from csbsju.edu to provide the most accurate and up-to-date schedule.", imageName: "clock")
+            InformationDetailView(title: "Accurate", subTitle: "Automatically pulls data from csbsju.edu to provide the most accurate and up-to-date schedule.", imageName: "clock")
 
-        InformationDetailView(title: "Made by a Johnnie", subTitle: "Made by a Johnnie, for my fellow Johnnies and Bennies.", imageName: "AppIcon"
-                              )
-    }
-    .padding(.horizontal)
+            InformationDetailView(title: "Made by a Johnnie", subTitle: "Made by a Johnnie, for my fellow Johnnies and Bennies.", imageName: "AppIcon"
+                                  )
+        }
+        .padding(.horizontal)
     }
 }
 }
