+++ Linkbus/Linkbus/SceneDelegate.swift
@@ -20,7 +20,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // Use a UIHostingController as window root view controller
         if let windowScene = scene as? UIWindowScene {
             let window = UIWindow(windowScene: windowScene)
-            window.rootViewController = UIHostingController(rootView: RouteList())
+            window.rootViewController = UIHostingController(rootView: Home())
             self.window = window
             window.makeKeyAndVisible()
         }
